test(frontend): icTransactions
